some updates

+++ src/mvs/driver.c
@@ -193,8 +193,8 @@ static void neogeo_set_cpu1_second_bank(UINT32 offset)
 	if (m68k_second_bank != offset)
 	{
 		m68k_second_bank = offset;
-		neogeo_cpu1_second_bank = (UINT16 *)((UINT32)memory_region_cpu1 + offset - 0x200000);
-		C68k_Set_Fetch(&C68K, 0x200000, 0x2fffff, (UINT32)&memory_region_cpu1[offset]);
+		neogeo_cpu1_second_bank = (UINT16 *)(memory_region_cpu1 + offset - 0x200000);
+		C68k_Set_Fetch(&C68K, 0x200000, 0x2fffff, &memory_region_cpu1[offset]);
 	}
 }
